Fix a bug with benchmarks against local dummy server failing with the latest exporter plugin version.
The root cause was the otel exporter plugin sending empty dictionary for
the `sessionInfo` body parameter. This can happen when sessionInfo is
empty because serverHost is sent as part of each event and not as part
of sessionInfo field.

Since tests only failed against local mock server and not against actual
DataSet prod US, this indicates the actual addEvents API indeed accepts
empty dictionary for the sessionInfo.

The mock server code was updated to accept empty sessionInfo object.

An alternative would be to update exporter plugin to not send that
attribute in the body in case it's empty, but that would be more complex
and may have more edge cases associated with it.